Improved the UI for adding a WiFi access point.
Before, each UI element needed to be modified, even if they were already correct.  Also, it was too easy to miss hitting the checkmark on the keyboard so that the password was never saved.  Now, the "default" values in the dropdowns are used, and if the user forgets to hit the checkmark, the watch puts up an error message and runs the UI again.