Added app to delete WiFi AP from memory.
Your stored WiFi access points (AP) are stored in a file in the SPIFF section of flash.
This app allows you to delete one of the AP in that file.  Useful if you make a mistake entering in a password for an AP, or if the password changes.  Just delete the AP and re-enter it.  This new app also needs changes to DudleyWatch.h, which I will upload separately.